completed the functionality using jquery in control.js
